Move concatUint8Arrays() helper to php-wasm/utils
We'll be reusing concatUint8Arrays() in more packages, e.g. php-wasm/universal. This
PR moves it into a globally available package like `@php-wasm/utils`

## Testing instructions

This is a small change. Just confirm the CI is green.
